Django_Rest_Framework/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
black==22.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
Deprecated==1.2.13
Django==4.0.2
django-admin-honeypot==1.1.0
django-filter==21.1
django-oauth-toolkit==1.7.0
django-redis==5.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
Faker==12.0.1
flake8==4.0.1
idna==3.3
importlib-metadata==4.11.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
jwcrypto==1.0
Markdown==3.3.6
MarkupSafe==2.1.0
mccabe==0.6.1
mixer==7.2.1
mypy-extensions==0.4.3
oauth2-provider==0.0
oauthlib==3.2.0
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
redis==4.1.4
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.8
wrapt==1.13.3
zipp==3.7.0
